pdr.Data: add __contains__, deprecate __iter__ (#57)
Add a __contains__ method to Data, which is True for all keys in
self.index and False otherwise.  This makes `if 'TABLE' in my_data`
return a useful value, mostly consistent with the dict-like behavior
of the existing __getitem__ method.

Deprecate the __iter__ method.  Iteration over a Data object was _not_
consistent with the behavior of __getitem__ (it iterated over values
rather than keys).  Worse, it forced loading of all the contained
data objects, which in the worst case could run the computer out of
memory.

Also deprecate __getitem__ being exactly equivalent to __getattribute__.
In a future release, only contained data objects will be considered
keys of the Data object, making [] indexing fully consistent with
the new __contains__ and the existing __len__ and keys methods.