Handle PHX-M-SSI-5-ATMOS-OPACITY-V1.0 format quirks.
Some of the table files in the PHX-M-SSI-5-ATMOS-OPACITY-V1.0 dataset
have lost trailing horizontal whitespace that their PDS3 label assumes
to be present.  This causes the default table parser to split them
into header and table at the wrong point, corrupting the parsed table.

Find the correct starting point for the table in these files by
counting bytes from the end of the file — this should be safe, since
the records of the actual table don’t have any trailing whitespace to
lose, and the record count in the label is correct.

Find the correct length of the header by counting its records as
newline-delimited instead of fixed-width — this should also be safe,
since these files are *both* fixed-width and newline-delimited, and
only the horizontal trailing whitespace was lost.  This change
exposed a bug in pdr.loaders.text.skeptically_load_header:
the hint argument to Python’s io.IOBase.readlines()
<https://docs.python.org/3.9/library/io.html#io.IOBase.readlines>
is the maximum number of *bytes* to read from the file, not the
maximum number of *lines*.  Change it to count lines itself.