address ndarray.newbyteorder() deprecation
numpy 2.0.0 removes the ndarray.newbyteorder() method. This change _should_ be nearly a
strict replacement except that it returns a view rather than a copy. Because we
specifically never do anything with the pre-enforced version of the array after
calling this method (why would we want to?) this should be safe, and a slight improvement
in terms of memory.