fixes/changes to multidict_dig_and_edit
multidict_dig_and_edit wasn't successfully digging into multidicts produced from some pds4 label elements. this fixes an incorrect conditional in that function and adds a capability to have 'predicate' check against the value rather than the key. note that this mandated a change in the signature of the 'predicate' argument.