CI: Merge coverage across all test jobs and generate HTML report.
Hat tip to https://hynek.me/articles/ditch-codecov-python/
for demonstrating how to do this entirely within Github Actions.

In order for it to be possible to merge coverage generated on a
Unixy platform with coverage generated on Windows, relative_files
mode is now enabled in `.coveragerc`.  This is unlikely to be an
issue for anyone but I thought it should be mentioned anyway.

TODO: Put the final HTML coverage report somewhere it can be
viewed on the web and doesn’t get deleted after 90 days.