Better way of monkey-patching a newer six.py into pds4_tools.
When patch_pds4_tools_six is called, use importlib to load our
known-good copy of six into the module path where pds4_tools will
expect it to be, superseding the bundled obsolete version.  For
future-proofing, only do this if pds4_tools fails to load without;
thus the patch will transparently cease to have any effect once
pds4_tools is fixed.

This eliminates the need to replace actual files inside the pds4_tools
package, which is brittle (e.g. it doesn’t work or even crashes if
pds4_tools isn’t to be found below `site.getsitepackages()[0]`), and
causes conda to complain about checksum mismatches,

To verify that this works, add a unit test for extremely basic parsing
of PDS4 labels.  Confirmed locally that this test succeeds with the
new patch in place and fails (actually, provokes an error during test
enumeration) using python 3.12 when the patch is stubbed out.