Fix CI failures with pip and python 3.9.
pyproject.toml conversion had the side effect of putting pytest in a
slightly different mode that causes it to trip over the outdated
pds4_tools.extern.six *whether or not* python 3.12 is in use.
(I do not yet know why it trips only when using pip to install
our dependencies, not conda, but it doesn’t really matter.)

While attempting to fix this I discovered *another* bug in pds4_tools:
it declares a dependency on tkinter only if the [viewer] feature is
active, but it loads part of tkinter unconditionally.  (This has also
already been fixed in the pds4_tools development tree.)

Fixed by:

 - generalizing patch_pds4_tools_six (now called patch_pds4_tools)
   to handle the lack of tkinter as well as the outdated copy of six
 - adding an argument to patch_pds4_tools to force installation of
   selected patches even if ‘import pds4_tools’ appears to work fine
 - calling patch_pds4_tools, with the force option, from conftest.py
   before test enumeration