CI: Run unit tests against installed wheels.
Instead of running the unit tests against an “editable install” of
the PDR git repo, build an sdist and wheel from the git repo, then,
in separate jobs that don’t ever check out the git repo, install the
wheel and run the tests from the sdist against the installed wheel.

This has several advantages.  Most importantly, it verifies that we
generate *correct* binary packages, as we were not doing until the
previous commit.  Testing an editable install is also subtly different
in terms of things like sys.path than testing something that’s been
officially installed to site-packages.[^1]  Finally, this puts us in
a better position to create *conda* packages (as a separate project).

It would be possible to add further automation that uploads the sdist
and wheel to PyPI if the commit has a release tag.

The mechanism for tweaking .coveragerc for CI’s needs has also been
overhauled, using a helper Python script that parses .coveragerc
exactly the same way coverage.py itself does (i.e. using configparser).
This should be much more reliable, both because we no longer need to
make assumptions about what goes where in the file, and because the
Python script can also use the coverage.py API to compute the path-
remapping configuration (yes, coverage.py ought to do this itself,
filed https://github.com/nedbat/coveragepy/issues/1840 ).  And it
means we can strip all the junk that’s in .coveragerc solely for
CI’s needs back out.

[^1]: https://setuptools.pypa.io/en/latest/userguide/development_mode.html#limitations