CI: tweaks for the sake of the coverage dashboard
Capture coverage data from each test job even if (some of) the tests
failed.  Store the merged coverage data as a coverage.py database
instead of a cobertura-format XML report.  Both these changes will
make it easier to produce a comprehensive coverage dashboard.

Also, rename all the intermediate CI artifacts so they appear at the
end of the alphabetical list and the persistent artifacts someone
might actually want to download by hand are at the top.  This doesn’t
have anything to do with the dashboard it was just bugging me to have
to scroll all the way to the bottom to get the “packages” artifact.