do not greedily head files to select attached labels
Previously, we always greedily read a megabyte from the head of every file that had (or might have) an attached label, then scanned through that megabyte of data to look for a label ending. This changes that behavior to iteratively scan through a file (up to the number of bytes we previously would have greedily read), stopping when and if we find an ending.

It also removes the 'fail silently by default for anything but a decode failure' behavior and adds a special-purpose Exception (InvalidAttachedLabel).