make-venv: stop trying to force use of “bin” on Windows
Having context.bin_path be “bin” on Windows breaks pip.  The only
reason we wanted this is so we didn’t have to write

    if [ -d venv/bin ]
    then source venv/bin/activate
    else source venv/Scripts/activate
    fi

all over the place in the workflow scripts, and we can achieve that
less invasively by creating venv/bin _as well as_ venv/Scripts and
putting copies of the activation scripts in there.